Candidates Master metadata

Basic information for each candidate.

candidates

Format

A data frame with 4648 rows and 15 columns, providing information for each candidate:

cand_id

A 9-character alpha-numeric code assigned to a candidate by the FEC, which remains the same across election cycles if running for the same office

cand_name

Candidate name

cand_pty_affiliation

The political party affiliation reported by the candidate

cand_election_yr

Candidate's Election Year

cand_office_st

Candidate's state

cand_office

Candidate office with designation: H = House, P = President, S = Senate

cand_office_district

Congressional district number: Congressional at-large 00, Senate 00, Presidential 00

cand_ici

Incumbent challenger status with designation: C = Challenger, I = Incumbent, O = Open Seat

cand_status

Candidate status with designation: C = Statutory candidate, F = Statutory candidate for future election, N = Not yet a statutory candidate, P = Statutory candidate in prior cycle

cand_pcc

ID of the candidate's principal campaign committee assigned by FEC

cand_st1

Candidate's Mailing address - street 1

cand_st2

Candidate's Mailing address - street 2

cand_city

Candidate's Mailing address - city

cand_st

Candidate's Mailing address - state

cand_zip

Candidate's Mailing address - ZIP code

Source

https://www.fec.gov/campaign-finance-data/candidate-master-file-description/, updated 2020-03-02