Committees metadata

Provides data pertaining to each committee registered with the Federal Election Commission.

committees

Format

A data frame with 17654 rows and 15 columns with information about FEC committees.

cmte_id

A 9-character alpha-numeric code assigned to a committee by the Federal Election Commission

cmte_nm

Name of the committee

tres_nm

The officially registered treasurer for the committee

cmte_st1

Street 1

cmte_st2

Street 2

cmte_city

City

cmte_st

State

cmte_zip

ZIP Code

cmte_dsgn

A = Authorized by a candidateB = Lobbyist/Registrant PACD = Leadership PACJ = Joint fundraiserP = Principal campaign committee of a candidateU = Unauthorized

cmte_tp

Committee Type, listed here: https://www.fec.gov/campaign-finance-data/committee-type-code-descriptions/

cmte_pty_affiliation

Committee Party, listed here: https://www.fec.gov/campaign-finance-data/party-code-descriptions/

cmte_filing_freq

Filing frequency: A = Administratively terminated,D = DebtM = Monthly filerQ = Quarterly filerT = TerminatedW = Waived

org_tp

Interest group category:C = CorporationL = Labor organizationM = Membership organizationT = Trade associationV = CooperativeW = Corporation without capital stock

connected_org_nm

Connected organization's name

cand_id

Candidate identification:When a committee has a committee type designation of H, S, or P, the candidate's identification number will be entered in this field.

Source

https://www.fec.gov/campaign-finance-data/committee-master-file-description/, updated 2020-03-02