Operating Expenditures

Operating expenditure information for a simple random sample of 1000 entries

expenditures

Format

A data frame with 1000 rows and 20 columns:

cmte_id

A 9-character alpha-numeric code assigned to a committee by the Federal Election Commission

amndt_ind

Amendment indicator: Indicates if the report being filed is new (N), an amendment (A) to a previous report or a termination (T) report

rpt_yr

Report year

rpt_tp

Indicates the type of report filed, listed here: https://www.fec.gov/campaign-finance-data/report-type-code-descriptions

line_num

Indicates FEC form line number

form_tp_cd

Indicates FEC form type

sched_tp_cd

Schedule type: Schedule B - Itemized disbursements

name

Contributor/lender/Transfer Name

city

City

state

State

zip_code

ZIP Code

transaction_dt

Transaction date (YYYY-MM-DD)

transaction_amt

Transaction Amount

transaction_pgi

Primary general indicator

purpose

Purpose

category

Disbursement category code: 001-012 and 101-107

category_desc

Disbursement Category Code Description, listed here: https://www.fec.gov/campaign-finance-data/disbursement-category-code-descriptions

entity_tp

Entity Type:CAN = CandidateCCM = Candidate CommitteeCOM = CommitteeIND = Individual (a person)ORG = Organization (not a committee and not a person)PAC = Political Action CommitteePTY = Party Organization

tran_id

Only for Electronic Filings. A unique identifier associated with each itemization or transaction appearing in an FEC electronic file. A transaction ID is unique for a specific committee for a specific report.

back_ref_tran_id

Back reference transaction ID: Used to associate one transaction with another transaction in the same report (using file number, transaction ID and back reference transaction ID). For example, a credit card payment and the sub-itemization of specific purchases. The back reference transaction ID of the specific purchases will equal the transaction ID of the payment to the credit card company.

Source

https://www.fec.gov/campaign-finance-data/operating-expenditures-file-description/, updated 2020-03-02