fec16: A Data Package for the 2016 Elections

The fec16 package provides the following 9 datasets from the Federal Election Commission's 2015-2016 election cycle:

Datasets

Candidates: candidates

Committees: committees

House Results: results_house

Senate Results: results_senate

Presidential Results: results_president

Individual Contributions: individuals

Committee Contributions: contributions

House/Senate Current Campaigns: campaigns

Operating Expenditures: expenditures

Transactions From One Committee To Another: transactions

Political Action Committee (PAC) and Party Summary Financial Information: pac

The entire candidates, committees, campaigns, results_house, results_senate, results_president, and pac datasets and samples containing 1000 entries of the other datasets are available directly in the package. For the datasets containing samples, full versions can be retrieved via the following functions:

Individual Contributions: read_all_individuals

Committee Contributions: read_all_contributions

Operating Expenditures: read_all_expenditures

Transactions From One Committee To Another: read_all_transactions

Authors

Marium Tapal: https://github.com/mariumtapal

Irene Ryan: https://github.com/ireneryan

Rana Gahwagy: https://github.com/ranawg

Ben Baumer: https://github.com/beanumber