Individual Contributions Master metadata

Information for a simple random sample of transactions from individuals to candidates/committees.

individuals

Format

A data frame with 1000 rows and 16 columns providing information for each transaction:

cmte_id

A 9-character alpha-numeric code assigned to a committee by the Federal Election Commission

amndt_ind

Amendment indicator: Indicates if the report being filed is new (N), an amendment (A) to a previous report or a termination (T) report

rpt_tp

Indicates the type of report filed, listed here: https://www.fec.gov/campaign-finance-data/report-type-code-descriptions/

transaction_pgi

This code indicates the election for which the contribution was made. EYYYY (election Primary, General, Other plus election year)

transaction_tp

Transaction types, listed here: https://www.fec.gov/campaign-finance-data/transaction-type-code-descriptions/

entity_tp

Entity Type:CAN = CandidateCCM = Candidate CommitteeCOM = CommitteeIND = Individual (a person)ORG = Organization (not a committee and not a person)PAC = Political Action CommitteePTY = Party Organization

name

Contributor/lender/Transfer Name

city

City

state

State

zip_code

ZIP Code

employer

Employer

occupation

Occupation

transaction_dt

Transaction date (YYYY-MM-DD)

transaction_amt

Transaction Amount

other_id

For contributions from individuals this column is null. For contributions from candidates or other committees this column will contain that contributor's FEC ID.

tran_id

Only for Electronic Filings. A unique identifier associated with each itemization or transaction appearing in an FEC electronic file. A transaction ID is unique for a specific committee for a specific report.

Source

https://www.fec.gov/campaign-finance-data/contributions-individuals-file-description/, updated 2020-03-02